Refactor FriendshipEventConsumer in notification-service
