+++ commons/src/main/java/com/eirsteir/coffeewithme/commons/domain/AbstractEntityNotificationEvent.java
@@ -1,6 +1,7 @@
 package com.eirsteir.coffeewithme.commons.domain;
 
 
+import io.eventuate.tram.events.common.DomainEvent;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.ToString;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public abstract class AbstractFriendshipEvent  implements FriendshipEvent {
+public abstract class AbstractEntityNotificationEvent implements DomainEvent {
 
     private Long subjectId;
     private UserDetails user;
