+++ commons/src/main/java/com/eirsteir/coffeewithme/commons/domain/FriendRequestAcceptedEvent.java
@@ -3,7 +3,7 @@ package com.eirsteir.coffeewithme.commons.domain;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
-public class FriendRequestAcceptedEvent extends AbstractFriendshipEvent {
+public class FriendRequestAcceptedEvent extends AbstractEntityNotificationEvent implements FriendshipEvent {
 
     public FriendRequestAcceptedEvent(Long subjectId, UserDetails user) {
         super(subjectId, user);
