+++ commons/src/main/java/com/eirsteir/coffeewithme/commons/domain/FriendRequestEvent.java
@@ -3,7 +3,7 @@ package com.eirsteir.coffeewithme.commons.domain;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
-public class FriendRequestEvent extends AbstractFriendshipEvent {
+public class FriendRequestEvent extends AbstractEntityNotificationEvent  implements FriendshipEvent{
 
     public FriendRequestEvent(Long subjectId, UserDetails user) {
         super(subjectId, user);
