+++ notification-service/src/main/java/com/eirsteir/coffeewithme/notification/service/NotificationService.java
@@ -1,8 +1,6 @@
 package com.eirsteir.coffeewithme.notification.service;
 
 
-import com.eirsteir.coffeewithme.notification.domain.Notification;
-import com.eirsteir.coffeewithme.commons.domain.NotificationType;
 import com.eirsteir.coffeewithme.notification.dto.NotificationDto;
 import org.springframework.data.domain.Pageable;
 
@@ -10,8 +8,6 @@ import java.util.List;
 
 
 public interface NotificationService {
-
-    void notify(Long subjectId, Notification.UserDetails currentUser, NotificationType type);
 
     NotificationDto updateNotificationToRead(NotificationDto notificationDto);
